Calculating Student Marks, Percentage, Semester Percentage, GPA, And Prevention of Deletions Semesters and Teachers When Subject is Related to them And Search By Kankor Id result is formatted.

+++ src/services/api/src/controllers/semester.controller.js
@@ -1,6 +1,6 @@
 const httpStatus = require('http-status');
 const catchAsync = require('../utils/catchAsync');
-const { semesterService, educationalYearService } = require('../services');
+const { semesterService, educationalYearService, subjectService } = require('../services');
 const ApiError = require('../utils/ApiError');
 
 const createSemester = catchAsync(async (req, res) => {
@@ -9,20 +9,22 @@ const createSemester = catchAsync(async (req, res) => {
   const semester = await semesterService.findSemester(req.body);
   if (semester) throw new ApiError(httpStatus.NOT_ACCEPTABLE, 'Semester is already created');
   const results = await semesterService.createNewSemester(req.body);
-  res.status(httpStatus.CREATED).send(results);
+  return res.status(httpStatus.CREATED).send(results);
 });
 
 const getSemester = catchAsync(async (req, res) => {
   const semester = await semesterService.findSemesterById(req.params.semesterId);
   if (!semester) throw new ApiError(httpStatus.NOT_FOUND, 'Semester Not Found');
-  res.status(httpStatus.OK).send(semester);
+  return res.status(httpStatus.OK).send(semester);
 });
 
 const deleteSemester = catchAsync(async (req, res) => {
-  const semester = await semesterService.findSemesterById(req.params.semesterId);
+  const semester = await semesterService.findById(req.params.semesterId);
   if (!semester) throw new ApiError(httpStatus.NOT_FOUND, 'semester not found');
+  const subjects = await subjectService.getSemesterStudents(req.params.semesterId);
+  if (subjects.length > 0) throw new ApiError(httpStatus.NOT_ACCEPTABLE, 'semester has subjects');
   await semesterService.deleteSemester(semester);
-  res.status(httpStatus.NO_CONTENT).send();
+  return res.status(httpStatus.NO_CONTENT).send();
 });
 
 const getSemesters = catchAsync(async (req, res) => {
@@ -33,7 +35,7 @@ const getSemesters = catchAsync(async (req, res) => {
     return res.status(httpStatus.OK).send(semesters);
   }
   const semesters = await semesterService.getAllSemesters();
-  res.status(httpStatus.OK).send(semesters);
+  return res.status(httpStatus.OK).send(semesters);
 });
